% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Box.R
\name{Box}
\alias{Box}
\title{Box}
\usage{
Box(
  children = list(),
  layout = Layout(),
  box_style = "",
  description = "Click Me",
  disabled = FALSE,
  ...,
  error_call = current_env()
)
}
\arguments{
\item{children}{children widgets}

\item{layout}{a \code{\link[=Layout]{Layout()}}}

\item{box_style}{box style. empty (default) or one of 'success', 'info', 'warning', 'danger', 'primary'.}

\item{description}{text description of the button}

\item{disabled}{TRUE if the Button is disabled}

\item{...}{These dots are for future extensions and must be empty.}

\item{error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
Box
}
